+++ c/clang.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-. ../scripts/profile.d/make_tl.sh
+. ../scripts/make_tl.sh
 
 # Default: gcc
 make_tl CC=clang "${@}"