+++ c/debug.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 set -e
 
-. ../scripts/profile.d/make_tl.sh
+. ../scripts/make_tl.sh
 make_tl DEBUG=1 "${@}"