+++ c/nopie.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-. ../scripts/profile.d/make_tl.sh
+. ../scripts/make_tl.sh
 
 make_tl clean
 make_tl NOPIE=1