+++ c/pie.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
-. ../scripts/profile.d/make_tl.sh
+. ../scripts/make_tl.sh
 make_tl clean
 make_tl PIE=1 "$@"